requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and Model Training
transformers>=4.35.0
accelerate>=0.24.0
peft>=0.7.0
datasets>=2.14.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
gensim>=4.3.0

# Visualization and Logging
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0

# Text Processing
sentencepiece>=0.1.99
tokenizers>=0.15.0

# Utilities
pyyaml>=6.0
safetensors>=0.4.0

# Optional: Word Cloud for Error Visualization
wordcloud>=1.9.0

# Optional: Jupyter for Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
